Activity test suite rewrote to Mocha
