+++ package.json
@@ -75,6 +75,7 @@
     "buffer": "^6.0.3",
     "chai": "^4.3.6",
     "chai-as-promised": "^7.1.1",
+    "chai-uuid": "^1.0.6",
     "cypress": "10.11.0",
     "eslint": "~8.22.0",
     "eslint-config-prettier": "^8.5.0",
